feat: set all analytics code as components
