+++ frontend/src/components/analytics/AnalyticsTable.tsx
